backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
minimalmodbus==2.1.1
pyserial==3.5
python-multipart==0.0.6
pydantic==2.5.3